refactor: convert runtime tests from external to internal package tests
Change test files from `runtime_test` to `runtime` package to enable
testing of package private functions. Remove ackrt import alias
as tests now have direct access to runtime package symbols.
